+++ policyfiles/policyfile.rb
@@ -1,0 +1,5 @@
+name 'backup'
+
+include_policy 'base', path: './base.lock.json'
+
+run_list 'managed-chef-server::policyfile_loader'