+++ recipes/backup.rb
@@ -3,16 +3,16 @@
 # Recipe:: backup
 #
 
-bdir = node['mcs']['backup']['dir']
-command = "#{bdir}/backup.sh"
+backupdir = node['mcs']['backup']['dir']
+command = "#{backupdir}/backup.sh"
 
-directory bdir
+directory backupdir
 
-# shell script for backups
+# shell script for backup
 file command do
   mode '0700'
   content "#/bin/sh
-cd #{bdir}
+cd #{backupdir}
 /opt/opscode/embedded/bin/knife ec backup --with-key-sql --with-user-sql -c /etc/opscode/pivotal.rb backup > backup.log 2>&1
 cd backup
 tar -czf ../#{node['mcs']['backup']['prefix']}`date +%Y%m%d%H%M`.tgz *
@@ -20,11 +20,11 @@ cd ..
 rm -rf backup"
 end
 
-# schedule backups on a recurring cron job. Refer to the README for further customization
+# schedule backup on a recurring cron job. Override attributes as necessary
 cron 'knife ec backup' do
-  environment ({ 'PWD' => bdir })
+  environment ({ 'PWD' => backupdir })
   command command
-  minute '*/5'
-  hour '*'
-  day '*'
+  minute node['mcs']['backup']['cron']['minute']
+  hour node['mcs']['backup']['cron']['hour']
+  day node['mcs']['backup']['cron']['day']
 end